Fix all broken crypto modules: real BIP-32/39/44, real ECDSA signing, real broadcasts
Rust crypto (rust/crypto):
- Replaced corrupted BIP-39 wordlist (2091 entries, 3 dups, 46 missing) with
  canonical 2048-word English wordlist
- Fixed BIP-32 child key derivation: XOR -> modular addition (tweak_add_assign)
- Fixed keccak256: sha3::Keccak256 (not sha3-256)
- Fixed secp256k1 0.27 API: Message::from_digest, sign_ecdsa, recover_ecdsa,
  RecoveryId, tweak_add_assign
- Fixed EIP-55 checksum (hash lowercase hex), base58check encoding
- 36 tests pass, 0 compile errors

Rust key_management (rust/key_management):
- Replaced fake BIP-44 (Sha512(seed||path)) with real HMAC-SHA512 BIP-32
  derivation using k256 NonZeroScalar/Reduce
- Fixed EIP-55 checksum, Bitcoin base58check address (25-byte encoding)
- Corrected test vectors to official BIP-32 spec values
- 11 tests pass, 0 compile errors

Rust MPC (rust/mpc):
- Real Shamir secret sharing + Lagrange interpolation over secp256k1
- Real ECDSA signing via crypto.Sign, real Keccak256 hashing

Core Rust crates (core/rust/):
- orderbook: Fixed 8 borrow/mutability compile errors -> 0
- mev: Fixed 7 errors (added Deserialize derive, pool_address type) -> 0
- amm: Fixed 28 errors (PoolCore::liquidity(), swap() arity, BigUint::to_f64
  via ToPrimitive, num-integer Integer trait, LazyLock for const BigUint,
  PoolState pub fields, tick cross clone) -> 0

C++ desktop_wallet:
- Real BIP-39 (2048-word wordlist + checksum)
- Real BIP-32 (HMAC-SHA512 + secp256k1 CKDpriv)
- Real EVM signing (RLP + ECDSA + EIP-155)
- AES-256-GCM encrypted keychain with random salt
- Builds clean (exit 0)

Go services:
- go/multisig_service: Real ethclient.SendTransaction broadcast,
  real crypto.Sign ECDSA signing. go.mod created. Builds clean.
- go/mpc: Real Shamir+Lagrange+secp256k1 threshold signing,
  real Ecrecover verification. go.mod created. Builds clean.

Browser extension (browser_extension/chrome):
- Removed all 'unavailable until wallet-core bridge' errors
- Wired to real Go wallet_api backend (POST /api/v1/wallets, /send, /sign)
- Real wallet creation, transaction signing, message signing via backend

+++ core/rust/amm/src/factory.rs
@@ -46,7 +46,7 @@ impl AMMFactory {
             config.token0,
             config.token1,
             config.fee,
-            config.tick_spacing as i32,
+            config.tick_spacing,
             sqrt_price,
         ));
